applied lead backend permissions

+++ src/app/api/Lead/delete/route.js
@@ -1,23 +1,22 @@
-import getDataFromToken from "@/helpers/getDataFromtoken";
-import { NextRequest, NextResponse } from "next/server";
-import Leads from "@/models/Leads";
 import connect from "@/dbConfig/dbConfig";
+import Leads from "@/models/Leads";
+import { NextResponse } from "next/server";
+import { checkPermission } from "../../permissions/checkPermission";
 
 connect();
 
 export async function DELETE(request) {
-    try {
-        const reqBody = await request.json();
-        const { leadIds } = reqBody;
-        await Leads.deleteMany({ _id: { $in: leadIds } })
-        return NextResponse.json({
-            mesaaage: "Leads found",
-        })
+  if (!(await checkPermission(request, "export", "lead")))
+    return NextResponse.json({ error: "Forbidden" }, { status: 403 });
 
-
-
-    }
-    catch (error) {
-        return NextResponse.json({ error: error.message }, { status: 400 })
-    }
+  try {
+    const reqBody = await request.json();
+    const { leadIds } = reqBody;
+    await Leads.deleteMany({ _id: { $in: leadIds } });
+    return NextResponse.json({
+      mesaaage: "Leads found",
+    });
+  } catch (error) {
+    return NextResponse.json({ error: error.message }, { status: 400 });
+  }
 }
